Refactor: User domain DTO 및 Repository 수정
