+++ src/auth/dto/auth.dto.ts
@@ -1,7 +1,15 @@
-import { IsOptional, IsString } from 'class-validator';
+import { IsEmail, IsInt, IsOptional, IsString } from 'class-validator';
 
 export class AuthResponse {
   @IsOptional()
   @IsString()
   accessToken: string;
+}
+
+export class DeserializeAccessToken {
+  @IsInt()
+  id: number;
+
+  @IsEmail()
+  email: string;
 }
