+++ src/auth/services/auth.service.ts
@@ -1,13 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { UserForAuth } from '../../user/vo/user.vo';
-import { AuthResponse } from '../dto/auth.dto';
+import { AuthResponse, DeserializeAccessToken } from '../dto/auth.dto';
 
 @Injectable()
 export class AuthService {
   constructor(private jwtService: JwtService) {}
 
-  login(user: UserForAuth): AuthResponse {
+  login(user: DeserializeAccessToken): AuthResponse {
     return {
       accessToken: this.jwtService.sign(user),
     };
