--- src/common/repositories/user.repository.ts
@@ -1,33 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { Prisma, User } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { CreateUserRequest } from '../../user/dto/user.dto';
-
-@Injectable()
-export class UserRepository {
-  constructor(private prisma: PrismaService) {}
-
-  async create(createUserRequest: CreateUserRequest): Promise<User> {
-    return this.prisma.user.create({
-      data: createUserRequest,
-    });
-  }
-
-  async findOne(
-    userWhereUniqueInput: Prisma.UserWhereUniqueInput,
-  ): Promise<User | null> {
-    return this.prisma.user.findUnique({
-      where: userWhereUniqueInput,
-    });
-  }
-
-  async updateOne(
-    userWhereUniqueInput: Prisma.UserWhereUniqueInput,
-    userUpdateInput: Prisma.UserUpdateInput,
-  ): Promise<User | null> {
-    return this.prisma.user.update({
-      where: userWhereUniqueInput,
-      data: userUpdateInput,
-    });
-  }
-}
