+++ src/user/controllers/user.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, Patch, UseGuards } from '@nestjs/common';
+import { DeserializeAccessToken } from 'src/auth/dto/auth.dto';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { User } from 'src/user/decorators/user.decorator';
 import { UpdateUserRequest, UpdateUserResponse } from '../dto/user.dto';
 import { UserService } from '../services/user.service';
-import { UserForAuth } from '../vo/user.vo';
 
 @Controller('user')
 export class UserController {
@@ -12,7 +12,7 @@ export class UserController {
   @UseGuards(JwtAuthGuard)
   @Patch('/')
   async update(
-    @User() user: UserForAuth,
+    @User() user: DeserializeAccessToken,
     @Body() updateUserRequest: UpdateUserRequest,
   ): Promise<UpdateUserResponse> {
     return this.userService.updateProfile(user, updateUserRequest);
