+++ src/user/services/user.service.ts
@@ -1,21 +1,16 @@
-import { Prisma } from '.prisma/client';
 import { Injectable } from '@nestjs/common';
-import { UserRepository } from 'src/common/repositories/user.repository';
+import { DeserializeAccessToken } from 'src/auth/dto/auth.dto';
 import { UpdateUserRequest, UpdateUserResponse } from '../dto/user.dto';
+import { UserRepository } from '../user.repository';
 
 @Injectable()
 export class UserService {
   constructor(private userRepository: UserRepository) {}
 
   updateProfile(
-    user: Prisma.UserWhereUniqueInput,
+    user: DeserializeAccessToken,
     updateUserRequest: UpdateUserRequest,
   ): Promise<UpdateUserResponse> {
-    return this.userRepository.updateOne(
-      {
-        email: user.email,
-      },
-      updateUserRequest,
-    );
+    return this.userRepository.updateByEmail(user, updateUserRequest);
   }
 }
