--- src/user/vo/user.vo.ts
@@ -1,12 +1,0 @@
-import { IsEmail, IsInt, IsString } from 'class-validator';
-
-export class UserForAuth {
-  @IsInt()
-  id: number;
-
-  @IsString()
-  name: string;
-
-  @IsEmail()
-  email: string;
-}
